ARM: Move in saparated file `data_processing`
